store-data-database

+++ crawler/spider/vnexpress.py
@@ -1,16 +1,16 @@
+from datetime import datetime
 import requests
 import sys
 from pathlib import Path
-from crawler.spider.connectsql import get_connection, insert_article
+from crawler.spider.connectsql import connect_to_sqlite, is_url_exists, insert_article
 from bs4 import BeautifulSoup
-
+import sqlite3
 FILE = Path(__file__).resolve()
 ROOT = FILE.parents[1]  # root directory
 if str(ROOT) not in sys.path:
     sys.path.append(str(ROOT))  # add ROOT to PATH
-
 from logger import log
-from crawler.spider.base_crawler import BaseCrawler
+from crawler.spider.base_crawler import Article, BaseCrawler
 from utils.bs4_utils import get_text_from_tag
 
 
@@ -41,11 +41,11 @@ class VNExpressCrawler(BaseCrawler):
 
     def extract_content(self, url: str) -> tuple:
         """
-        Extract title, description and paragraphs from url
+        Extract title, author and content from url
         @param url (str): url to crawl
         @return title (str)
-        @return description (generator)
-        @return paragraphs (generator)
+        @return author (generator)
+        @return content (generator)
         """
         content = requests.get(url).content
         soup = BeautifulSoup(content, "html.parser")
@@ -53,40 +53,42 @@ class VNExpressCrawler(BaseCrawler):
         title = soup.find("h1", class_="title-detail") 
         if title == None:
             return None, None, None
+        
         title = title.text
 
-        # some sport news have location-stamp child tag inside description tag
-        description = (get_text_from_tag(p) for p in soup.find("p", class_="description").contents)
-        paragraphs = (get_text_from_tag(p) for p in soup.find_all("p", class_="Normal"))
+        # some sport news have location-stamp child tag inside author tag
+        author = (get_text_from_tag(p) for p in soup.find("p", class_="description").contents)
+        content = (get_text_from_tag(p) for p in soup.find_all("p", class_="Normal"))
 
-        return title, description, paragraphs
+        return title, author, content
 
     def write_content(self, url: str, output_fpath: str) -> bool:
         """
-        From url, extract title, description and paragraphs then write in output_fpath
+        From url, extract title, author and content then write in output_fpath
         @param url (str): url to crawl
         @param output_fpath (str): file path to save crawled result
         @return (bool): True if crawl successfully and otherwise
         """
-        title, description, paragraphs = self.extract_content(url)
+        title, author, content = self.extract_content(url)
 
-        if title == None:
-            return False
-
-        with open(output_fpath, "w", encoding="utf-8") as file:
-            file.write(title + "\n")
-            for p in description:
-                file.write(p + "\n")
-            for p in paragraphs:                     
-                file.write(p + "\n")
         
-        article = Article(title, '\n'.join(description),
-                          '\n'.join(paragraphs), url)
-
+        
+        article = Article(id=None,title='\n'.join(title), url=url, image_url="", author='\n'.join(
+            author), content='\n'.join(content), created_at=datetime.now(), sentiment="", is_fake=False)
+        connect_sql = connect_to_sqlite(database_file="osintnews.db")
+        
+        
+        if not is_url_exists(connect_sql, article.url):
+            if title == None:
+                return False
+            with open(output_fpath, "w", encoding="utf-8") as file:
+                file.write(title + "\n")
+                for p in author:
+                    file.write(p + "\n")
+                for p in content:                     
+                    file.write(p + "\n")
         # Insert the article into the database
-        insert_article(article)
-        get_connection()
-        
+            insert_article(connect_sql, article)
         
         return True
